+++ src/sessions/softphone-session-handler.ts
@@ -392,6 +392,9 @@ export class SoftphoneSessionHandler extends BaseSessionHandler {
     callState: ICallStateFromParticipant
   ): void {
     if (callState.errorInfo) {
+      if (callState.errorInfo.code === 'error.ininedgecontrol.session.inactive') {
+        return;
+      }
       this.debouncedEmitCallError(update, participant, callState);
     }
   }